Move Deposit to the types crate
And this commit renames another Deposit type which means the quantity of
deposit to DepositQuantity.
