+++ core/src/consensus/tendermint/mod.rs
@@ -27,7 +27,7 @@ mod worker;
 
 use self::chain_notify::TendermintChainNotify;
 pub use self::message::{ConsensusMessage, VoteOn, VoteStep};
-pub use self::params::{Deposit, TendermintParams, TimeGapParams, TimeoutParams};
+pub use self::params::{TendermintParams, TimeGapParams, TimeoutParams};
 pub use self::types::{Height, Step, View};
 pub use super::{stake, ValidatorSet};
 use crate::client::ConsensusClient;
@@ -38,6 +38,7 @@ use crate::ChainNotify;
 use ckey::Address;
 use crossbeam_channel as crossbeam;
 use ctimer::TimerToken;
+use ctypes::Deposit;
 use parking_lot::RwLock;
 use std::collections::HashMap;
 use std::sync::atomic::AtomicBool;
