+++ types/src/deposit.rs
@@ -1,4 +1,4 @@
-// Copyright 2018-2020 Kodebox, Inc.
+// Copyright 2020 Kodebox, Inc.
 // This file is part of CodeChain.
 //
 // This program is free software: you can redistribute it and/or modify
@@ -14,31 +14,13 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-#[macro_use]
-extern crate serde_derive;
-#[macro_use]
-extern crate rlp_derive;
+use ckey::Ed25519Public as Public;
+use primitives::Bytes;
 
-mod block_hash;
-mod common_params;
-mod deposit;
-mod tracker;
-mod tx_hash;
-mod validator_set;
-
-pub mod errors;
-pub mod header;
-pub mod transaction;
-pub mod util;
-
-pub type BlockNumber = u64;
-pub type ShardId = u16;
-
-pub use block_hash::BlockHash;
-pub use common_params::CommonParams;
-pub use deposit::Deposit;
-pub use header::Header;
-pub use tracker::Tracker;
-pub use tx_hash::TxHash;
-pub use validator_set::CompactValidatorEntry;
-pub use validator_set::CompactValidatorSet;
+#[derive(Clone, Eq, Ord, PartialEq, PartialOrd)]
+pub struct Deposit {
+    pub pubkey: Public,
+    pub deposit: u64,
+    pub nomination_ends_at: u64,
+    pub metadata: Bytes,
+}
